Splitting out dask work from multi -> dask_processing - Added more documentation
